ML/Agent/shell/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
transformers>=4.36.0
torch>=2.1.0  # Required for transformers